+++ NEWS.md
@@ -1,7 +1,9 @@
 # BrainGnomes 0.7-3
 
 * bugfix: correctly handle unsigned integer data types in NIfTIs
-* lmfit_residuals_4d now handles partial (ala fsl_regfilt) and full regression
+* lmfit_residuals_4d now handles partial (ala fsl_regfilt) and full regression and is used for applying AROMA
+* Nonaggressive and aggressive AROMA now supported
+* fsl_regfilt.R wrapper script removed from pipeline -- all regression now happens with lmfit_residuals_4d
 
 # BrainGnomes 0.7-2